requirements.txt
flask
flask-cors
openai
pandas
matplotlib
seaborn
mysql-connector-python
python-dotenv
gunicorn